requirements.txt
Django==1.11.2
django-cors-headers==2.1.0
django-debug-toolbar==1.8
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
PyJWT==1.5.0
pytz==2017.2
sqlparse==0.2.3
